influxdb_writer/requirements.txt
influxdb-client==1.49.0
paho-mqtt==2.1.0
